Database Inclusion
SQLlite database hooked up.
Data serialisation and persistence complete.
Need to now do this for all schemas, only currently complete for Professors

I should really include file changes here but am pushed for time this morning, but the jist is I have used SQLAlchemy as an ORM and Marshmallow to serialise the data